Review problems form 35 to 51

+++ course 7 - mysolution/problem #46.cpp
@@ -27,7 +27,7 @@ sClient ConvertLineToRecord(string stLine, string Seperator = "#//#")
 	Client.PinCode = vClientData[1];
 	Client.Name = vClientData[2];
 	Client.Phone = vClientData[3];
-	Client.AccountBalance = stod(vClientData[4]);
+	Client.AccountBalance = stoi(vClientData[4]);
 	
 	return Client;
 }
